+++ api/commands/plotman_cli.py
@@ -239,10 +239,10 @@ def analyze(plot_file):
         except TimeoutExpired:
             proc.kill()
             proc.communicate()
-            abort(500, description="The timeout is expired attempting to start plots check.")
+            abort(500, description="The timeout is expired attempting to start plot analyze.")
         if errs:
             app.logger.error(errs.decode('utf-8'))
-            abort(500, description="Failed to analyze plots.")
+            abort(500, description="Failed to analyze plot.")
         return outs.decode('utf-8')
     return None