+++ api/views/certificates/resources.py
@@ -51,6 +51,9 @@ class Certificates(MethodView):
         elif type == "hddcoin":
             blockchain = "hddcoin"
             dir = "/root/.hddcoin/mainnet/config/ssl/ca"
+        elif type == "maize":
+            blockchain = "maize"
+            dir = "/root/.maize/mainnet/config/ssl/ca"
         elif type == "silicoin":
             blockchain = "silicoin"
             dir = "/root/.silicoin/mainnet/config/ssl/ca"