+++ common/models/plots.py
@@ -17,5 +17,7 @@ class Plot(db.Model):
     dir = sa.Column(sa.String(length=255), nullable=False)
     file = sa.Column(sa.String(length=255), nullable=False)
     size = sa.Column(sa.Integer, nullable=False)
+    plot_check = sa.Column(sa.String(length=255), nullable=True)
+    plot_analyze = sa.Column(sa.String(length=255), nullable=True)
     created_at = sa.Column(sa.String(length=64), nullable=False)
     updated_at = sa.Column(sa.DateTime(), onupdate=func.now())