+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.6] - 2021-11-28
+ - [Maize](https://github.com/Maize-Network/maize-blockchain) - cross-farming support for this blockchain fork.
+ - [Forktools](https://github.com/Qwinn1/forktools) - utilities including a multiproc patch to limit memory usage.
+ - Farming - right-side of Plots table offers Chia "plots check", slowly generated on background thread.
+ - Bug Fix - hourly log rotation for `farmr`, auto-deletion log if above 20 MB in size.  Sorry all!
+
 ## [0.6.5] - 2021-11-19
  - [Cryptodoge](https://github.com/CryptoDoge-Network/cryptodoge) - cross-farming support for this blockchain fork.
  - Docker images now roughly 1/3 the size of previous releases.  Shared base image further decreases download size for forks.